fix tests for real this time

+++ pkg/services/publicdashboards/service/query_test.go
@@ -890,7 +890,7 @@ func TestBuildMetricRequest(t *testing.T) {
 				},
 				"intervalMs":      int64(10000000),
 				"maxDataPoints":   int64(200),
-				"queryCachingTTL": 0,
+				"queryCachingTTL": int64(0),
 				"refId":           "A",
 			}),
 			reqDTO.Queries[0],
